postgres: widen event-stream sequence columns to BIGINT
flue_event_streams.next_offset and flue_event_stream_entries.seq were
32-bit INTEGER while sibling counters in the same adapter use BIGINT
and the SQLite backend stores 64-bit integers. Reads already wrap these
columns in Number(...), so the DDL change needs no code changes.

Review finding: F243 (plans/2026-06-11-pre-beta-full-review-findings.md)

+++ packages/postgres/src/postgres-adapter.ts
@@ -353,7 +353,7 @@ async function ensureTables(runner: PgRunner): Promise<void> {
 		await tx.query(`
 			CREATE TABLE IF NOT EXISTS flue_event_streams (
 				path         TEXT PRIMARY KEY,
-				next_offset  INTEGER NOT NULL DEFAULT 0,
+				next_offset  BIGINT NOT NULL DEFAULT 0,
 				closed       BOOLEAN NOT NULL DEFAULT FALSE
 			)
 		`);
@@ -361,7 +361,7 @@ async function ensureTables(runner: PgRunner): Promise<void> {
 		await tx.query(`
 			CREATE TABLE IF NOT EXISTS flue_event_stream_entries (
 				path    TEXT NOT NULL,
-				seq     INTEGER NOT NULL,
+				seq     BIGINT NOT NULL,
 				data    TEXT NOT NULL,
 				PRIMARY KEY (path, seq)
 			)
